Updating travis
[noissue]

+++ .travis/before_install.sh
@@ -19,13 +19,20 @@ if [ -f $PRE_BEFORE_INSTALL ]; then
     $PRE_BEFORE_INSTALL
 fi
 
+if [[ -n $(echo -e $COMMIT_MSG | grep -P "Required PR:.*" | grep -v "https") ]]; then
+  echo "Invalid Required PR link detected in commit message. Please use the full https url."
+  exit 1
+fi
+
 export PULP_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/pulp\/pulpcore\/pull\/(\d+)' | awk -F'/' '{print $7}')
 export PULP_PLUGIN_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/pulp\/pulpcore-plugin\/pull\/(\d+)' | awk -F'/' '{print $7}')
 export PULP_SMASH_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/PulpQE\/pulp-smash\/pull\/(\d+)' | awk -F'/' '{print $7}')
 export PULP_ROLES_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/pulp\/ansible-pulp\/pull\/(\d+)' | awk -F'/' '{print $7}')
 export PULP_BINDINGS_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/pulp\/pulp-openapi-generator\/pull\/(\d+)' | awk -F'/' '{print $7}')
+export PULP_OPERATOR_PR_NUMBER=$(echo $COMMIT_MSG | grep -oP 'Required\ PR:\ https\:\/\/github\.com\/pulp\/pulp-operator\/pull\/(\d+)' | awk -F'/' '{print $7}')
 
-# dev_requirements should not be needed for testing; don't install them to make sure
+# test_requirements contains tools needed for flake8, etc.
+# So install them here rather than in install.sh
 pip install -r test_requirements.txt
 
 # check the commit message
@@ -39,55 +46,61 @@ flake8 --config flake8.cfg
 
 cd ..
 git clone --depth=1 https://github.com/pulp/ansible-pulp.git
-if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
-  if [ -n "$PULP_ROLES_PR_NUMBER" ]; then
-    cd ansible-pulp
-    git fetch --depth=1 origin +refs/pull/$PULP_ROLES_PR_NUMBER/merge
-    git checkout FETCH_HEAD
-    cd ..
-  fi
+if [ -n "$PULP_ROLES_PR_NUMBER" ]; then
+  cd ansible-pulp
+  git fetch --depth=1 origin +refs/pull/$PULP_ROLES_PR_NUMBER/merge
+  git checkout FETCH_HEAD
+  cd ..
 fi
+
+
+git clone --depth=1 https://github.com/pulp/pulp-operator.git
+if [ -n "$PULP_OPERATOR_PR_NUMBER" ]; then
+  cd pulp-operator
+  git fetch --depth=1 origin +refs/pull/$PULP_OPERATOR_PR_NUMBER/merge
+  git checkout FETCH_HEAD
+  cd ..
+fi
+
 
 git clone --depth=1 https://github.com/pulp/pulpcore.git
 
-if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
-  if [ -n "$PULP_PR_NUMBER" ]; then
-    cd pulpcore
-    git fetch --depth=1 origin +refs/pull/$PULP_PR_NUMBER/merge
-    git checkout FETCH_HEAD
-    cd ..
-  fi
+if [ -n "$PULP_PR_NUMBER" ]; then
+  cd pulpcore
+  git fetch --depth=1 origin +refs/pull/$PULP_PR_NUMBER/merge
+  git checkout FETCH_HEAD
+  cd ..
 fi
 
 
-git clone --depth=1 https://github.com/pulp/pulpcore-plugin.git
+# When building a (release) tag, we don't need the development modules for the
+# build (they will be installed as dependencies of the plugin).
+if [ -z "$TRAVIS_TAG" ]; then
 
-if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
+  git clone --depth=1 https://github.com/pulp/pulpcore-plugin.git
+
   if [ -n "$PULP_PLUGIN_PR_NUMBER" ]; then
     cd pulpcore-plugin
     git fetch --depth=1 origin +refs/pull/$PULP_PLUGIN_PR_NUMBER/merge
     git checkout FETCH_HEAD
     cd ..
   fi
-fi
 
 
-git clone --depth=1 https://github.com/PulpQE/pulp-smash.git
+  git clone --depth=1 https://github.com/PulpQE/pulp-smash.git
 
-if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
   if [ -n "$PULP_SMASH_PR_NUMBER" ]; then
     cd pulp-smash
     git fetch --depth=1 origin +refs/pull/$PULP_SMASH_PR_NUMBER/merge
     git checkout FETCH_HEAD
     cd ..
   fi
+
+  # pulp-smash already got installed via test_requirements.txt
+  pip install --upgrade --force-reinstall ./pulp-smash
 fi
 
-psql -c 'CREATE DATABASE pulp OWNER travis;'
-
 pip install ansible
-cp pulp_ansible/.travis/playbook.yml ansible-pulp/playbook.yml
-cp pulp_ansible/.travis/postgres.yml ansible-pulp/postgres.yml
 
 cd pulp_ansible
